+++ package/mirage-lc/src/Mirage/Domain/Setting.fs
@@ -14,18 +14,12 @@ open Mirage.PluginInfo
 type Settings =
     {   /// Volume used for voice play-back on monsters mimicking the local player. Must be a value between 0.0f-1.0f
         localPlayerVolume: float32
-        /// Whether monsters mimicking the local player should be muted while the player is alive.
-        hearLocalVoiceWhileAlive: bool
-        /// Whether the microphone should continue recording while the local player is dead.
-        recordWhileDead: bool
         /// If set to false, recordings are deleted when closing the game by default.
         neverDeleteRecordings: bool
     }
 
 let defaultSettings =
     {   localPlayerVolume = 1.0f
-        hearLocalVoiceWhileAlive = true
-        recordWhileDead = false
         neverDeleteRecordings = false
     }
 
@@ -68,16 +62,6 @@ let initSettings filePath =
                         MaxValue = 100.0f,
                         Text = "Volume when a monster is mimicking your own voice:",
                         OnValueChanged = fun _ value -> saveSettings { settings with localPlayerVolume = value / 100.0f }
-                    )
-                    ToggleComponent(
-                        Text = "Only hear a monster mimicking your own voice while spectating",
-                        Value = not settings.hearLocalVoiceWhileAlive,
-                        OnValueChanged = fun _ value -> saveSettings { settings with hearLocalVoiceWhileAlive = not value }
-                    )
-                    ToggleComponent(
-                        Text = "Only record your voice while alive",
-                        Value = not settings.recordWhileDead,
-                        OnValueChanged = fun _ value -> saveSettings { settings with recordWhileDead = not value }
                     )
                     ToggleComponent(
                         Text = "Never delete recordings",