+++ package/mirage-lc/src/Mirage/Plugin.fs
@@ -8,6 +8,7 @@ open System.Diagnostics
 open NAudio.Lame
 open UnityEngine
 open Mirage.PluginInfo
+open Mirage.Compatibility
 open Mirage.Domain.Netcode
 open Mirage.Domain.Logger
 open Mirage.Domain.Setting
@@ -20,6 +21,7 @@ open Mirage.Hook.Dissonance
 open Mirage.Hook.MaskedPlayerEnemy
 
 [<BepInPlugin(pluginId, pluginName, pluginVersion)>]
+[<BepInDependency(LobbyCompatibility.PluginInfo.PLUGIN_GUID, BepInDependency.DependencyFlags.SoftDependency)>]
 type Plugin() =
     inherit BaseUnityPlugin()
 
@@ -40,6 +42,7 @@ type Plugin() =
         for category in Seq.cast<LogCategory> <| Enum.GetValues typeof<LogCategory> do
             Logs.SetLogLevel(category, LogLevel.Error)
 
+        initLobbyCompatibility()
         initRecordingManager recordingDirectory
         initSettings <| Path.Join(mirageDirectory, "settings.json")
         initNetcodePatcher()